Moved services in their right location. Now the solution is cleared and easier to be understood
